+++ BambooTracker/gui/effect_list_dialog.cpp
@@ -70,6 +70,7 @@ EffectListDialog::EffectListDialog(QWidget *parent) :
 	addRow(EffectType::ARControl, { SoundSource::FM });
 	addRow(EffectType::Brightness, { SoundSource::FM });
 	addRow(EffectType::DRControl, { SoundSource::FM });
+	addRow(EffectType::EnvelopeReset, { SoundSource::FM });
 	addRow(EffectType::FBControl, { SoundSource::FM });
 	addRow(EffectType::FineDetune, { SoundSource::FM, SoundSource::SSG, SoundSource::ADPCM });
 	addRow(EffectType::VolumeDelay, { SoundSource::FM, SoundSource::SSG, SoundSource::RHYTHM, SoundSource::ADPCM });